requirements.txt
requests==2.28.1
python-dotenv==0.21.0
apscheduler==3.9.1
tweepy==4.10.1
boto3==1.24.89